requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
psycopg==3.2.5
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
yarl==1.18.3
